SE-342/assignments/ass2/requirements.txt
numpy==1.17.4
scipy==1.3.2
opencv_python==4.1.2.30
Pillow==6.2.1
PyQt5==5.13.2
PySide2==5.13.2
